+++ paimon-core/src/main/java/org/apache/paimon/table/format/FormatTableCommitMessageSerializer.java
@@ -16,23 +16,18 @@
  * limitations under the License.
  */
 
-package org.apache.paimon.table.sink;
+package org.apache.paimon.table.format;
 
 import org.apache.paimon.annotation.Public;
 import org.apache.paimon.fs.Path;
-import org.apache.paimon.table.sink.FormatTableAtomicCommitter.TempFileInfo;
-import org.apache.paimon.table.sink.FormatTableWrite.FormatTableCommitMessage;
+import org.apache.paimon.table.format.FormatTableAtomicCommitter.TempFileInfo;
+import org.apache.paimon.table.format.FormatTableWrite.FormatTableCommitMessage;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 
-/**
- * Serializer for {@link FormatTableCommitMessage} that provides consistent serialization across
- * distributed environments following Spark/Flink patterns.
- *
- * @since 0.9.0
- */
+/** Serializer for {@link FormatTableCommitMessage} that provides consistent serialization. */
 @Public
 public class FormatTableCommitMessageSerializer {
