+++ paimon-core/src/test/java/org/apache/paimon/table/sink/FormatTableAtomicWriteTest.java
@@ -26,8 +26,9 @@ import org.apache.paimon.fs.FileStatus;
 import org.apache.paimon.fs.Path;
 import org.apache.paimon.fs.local.LocalFileIO;
 import org.apache.paimon.table.FormatTable;
-import org.apache.paimon.table.sink.FormatTableAtomicCommitter.TempFileInfo;
-import org.apache.paimon.table.sink.FormatTableWrite.FormatTableCommitMessage;
+import org.apache.paimon.table.format.FormatTableAtomicCommitter;
+import org.apache.paimon.table.format.FormatTableAtomicCommitter.TempFileInfo;
+import org.apache.paimon.table.format.FormatTableWrite.FormatTableCommitMessage;
 import org.apache.paimon.types.DataType;
 import org.apache.paimon.types.DoubleType;
 import org.apache.paimon.types.IntType;
